\name{SC19051-package}
\alias{SC19051-package}
\alias{SC19051}
\docType{package}
\title{
  Clustering in the gaussian mixture model
}
\description{
  Gaussian mixture model is a classic mixture model and the EM algorithm has been successfully applied. We apply another method in this package: constrained Newton method, which we can refer to [Wang, 2007]
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Sirui Lin, 
Maintainer: Sirui Lin <lsr@mail.ustc.edu.cn>
}
\references{
  Wang, 2007, _On fast computation of the non-parametricmaximum likelihood estimate of a mixing distribution_
 
}
\keyword{ package }

